explicitly cast some variable because msvc is dumb
svn path=/trunk/extragear/network/ktorrent/; revision=824462

+++ ktorrent/recommendedsettingsdlg.cpp
@@ -109,7 +109,7 @@ namespace kt
 		qreal avg_slot_up = ceil(qMax(pow(upload_rate/3.5,0.55),4.0));
 		
 		Uint32 max_torrents = qRound(pow(upload_rate * 0.25,0.3));
-		max_downloads = ceil(max_torrents * 2/3);
+		max_downloads = ceil((float)(max_torrents * 2/3));
 		max_seeds = qMax(max_torrents - max_downloads,(bt::Uint32)1);
 		
 		if (m_chk_avg_speed_slot->isChecked())
@@ -135,26 +135,26 @@ namespace kt
 		else if (m_chk_avg_speed_slot->isChecked() && m_chk_slots->isChecked())
 		{
 			max_torrents = qRound(max_upload_speed / (avg_slot_up*max_slots));
-			max_downloads = ceil(max_torrents * 2/3);
+			max_downloads = ceil((float)(max_torrents * 2/3));
 			max_seeds = qMax(max_torrents - max_downloads,(bt::Uint32)1);
 		}
 		else if (m_chk_sim_torrents->isChecked() && m_chk_slots->isChecked())
 		{
-			avg_slot_up = ceil(max_upload_speed / (max_slots * max_torrents));
+			avg_slot_up = ceil((float)(max_upload_speed / (max_slots * max_torrents)));
 		}
 		else if (m_chk_slots->isChecked())
 		{
 			avg_slot_up = ceil(qMax(pow(max_upload_speed/3.5,0.55),4.0)); // basis to calculate the number of torrents
 			max_torrents = qRound(max_upload_speed / (avg_slot_up*max_slots));
-			max_downloads = ceil(max_torrents * 2/3);
+			max_downloads = ceil((float)(max_torrents * 2/3));
 			max_seeds = qMax(max_torrents - max_downloads,(bt::Uint32)1);
-			avg_slot_up = ceil(max_upload_speed/(max_slots*max_torrents)); // real number after the slots have been multiplied with the torrents
+			avg_slot_up = ceil((float)(max_upload_speed/(max_slots*max_torrents))); // real number after the slots have been multiplied with the torrents
 		}
 		
 		if (max_downloads == 0)
 			max_downloads = 1;
 		
-		max_conn_glob = qRound(qMin(pow(upload_rate*8,0.8)+50,900.0));
+		max_conn_glob = qRound(qMin((double)pow((int)(upload_rate*8),0.8)+50,900.0));
 		max_conn_tor = qRound(qMin(max_conn_glob * 1.2 / max_torrents,(qreal)max_conn_glob));
 	
 		m_max_upload->setText(QString("<b>%1</b>").arg(KBytesPerSecToString(max_upload_speed)));
